\name{helloR-package}
\alias{helloR-package}
\alias{helloR}
\docType{package}
\title{
\packageTitle{helloR}
}
\description{
\packageDescription{helloR}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{helloR}

Maintainer: \packageMaintainer{helloR}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = helloR)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{helloR}}
%\section{Documentation Index}{\packageIndices{helloR}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
